pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dl-app-template"
version = "0.1.0"
description = "A deep learning app template with Streamlit frontend and FastAPI backend"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn[standard]>=0.15.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "streamlit>=1.0.0",
    "sqlalchemy>=1.4.0",
    "pymysql>=1.0.0",
    "pillow>=8.0.0",
    "python-multipart>=0.0.5",
    "python-dotenv>=0.19.0",
    "rich>=10.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "app.*"]